Time: 1 ms (96.61%) | Memory: 42.3 MB (94.38%) - LeetSync

+++ 231-power-of-two/power-of-two.java
@@ -1,7 +1,7 @@
 class Solution {
     public boolean isPowerOfTwo(int n) {
         if(n<=0)return false;
-        long result=n&n-1;
+        int result=n&n-1;
         if(result==0)
         return true;
         else
